moved more mapping classes into their own files and passed configuration options through the constructors
