+++ CMakeLists.txt
@@ -103,6 +103,8 @@ set(LIB_SRC_FILES
     ${SRC_DIR}/nektar_interface/particle_cell_mapping/map_particles_2d_regular.cpp
     ${SRC_DIR}/nektar_interface/particle_cell_mapping/map_particles_3d.cpp
     ${SRC_DIR}/nektar_interface/particle_cell_mapping/map_particles_3d_regular.cpp
+    ${SRC_DIR}/nektar_interface/particle_cell_mapping/map_particles_common.cpp
+    ${SRC_DIR}/nektar_interface/particle_cell_mapping/map_particles_host.cpp
     ${SRC_DIR}/plasma.cpp
     ${SRC_DIR}/run_info.cpp
     ${SRC_DIR}/simulation.cpp
@@ -150,6 +152,8 @@ set(HEADER_FILES
     ${INC_DIR}/nektar_interface/particle_cell_mapping/map_particles_2d_regular.hpp
     ${INC_DIR}/nektar_interface/particle_cell_mapping/map_particles_3d.hpp
     ${INC_DIR}/nektar_interface/particle_cell_mapping/map_particles_3d_regular.hpp
+    ${INC_DIR}/nektar_interface/particle_cell_mapping/map_particles_common.hpp
+    ${INC_DIR}/nektar_interface/particle_cell_mapping/map_particles_host.hpp
     ${INC_DIR}/nektar_interface/particle_cell_mapping/map_particles_newton.hpp
     ${INC_DIR}/nektar_interface/particle_cell_mapping/mapping_newton_iteration_base.hpp
     ${INC_DIR}/nektar_interface/particle_cell_mapping/newton_geom_interfaces.hpp
