+++ include/nektar_interface/parameter_store.hpp
@@ -34,6 +34,16 @@ public:
   }
 
   /**
+   *  Test if key exists in store.
+   *
+   *  @param name Key to test for.
+   *  @returns True if key exists in store.
+   */
+  inline bool contains(const std::string name) {
+    return (bool)this->store.count(name);
+  }
+
+  /**
    *  Retrieve a REAL or INT value stored against a given key. If the key is not
    *  found return the passed value as a default.
    *
@@ -45,7 +55,7 @@ public:
   inline T get(const std::string name, const T value = 0) {
     static_assert((std::is_same_v<T, REAL> || std::is_same_v<T, INT>),
                   "Bad type passed to ParameterStore::get");
-    if (this->store.count(name)) {
+    if (this->contains(name)) {
       return std::get<T>(this->store[name]);
     } else {
       return value;
