+++ include/nektar_interface/particle_cell_mapping/map_particles_host.hpp
@@ -1,5 +1,5 @@
-#ifndef __MAP_PARTICLES_2D_H__
-#define __MAP_PARTICLES_2D_H__
+#ifndef __MAP_PARTICLES_HOST_H__
+#define __MAP_PARTICLES_HOST_H__
 
 #include <cmath>
 #include <map>
@@ -7,18 +7,13 @@
 #include <mpi.h>
 #include <vector>
 
-#include "coarse_lookup_map.hpp"
-#include "map_particles_2d_regular.hpp"
-#include "nektar_interface/coordinate_mapping.hpp"
-#include "nektar_interface/geometry_transport/shape_mapping.hpp"
+#include "../particle_mesh_interface.hpp"
 #include "nektar_interface/parameter_store.hpp"
-#include "nektar_interface/particle_mesh_interface.hpp"
-#include "newton_geom_interfaces.hpp"
 #include "particle_cell_mapping_common.hpp"
-
 #include <SpatialDomains/MeshGraph.h>
 #include <neso_particles.hpp>
 
+using namespace Nektar;
 using namespace Nektar::SpatialDomains;
 using namespace NESO;
 using namespace NESO::Particles;
@@ -26,25 +21,25 @@ using namespace NESO::Particles;
 namespace NESO {
 
 /**
- *  Class to map particles into Nektar++ cells.
+ *  Class to map particles into Nektar++ cells on the CPU host. Should work for
+ *  all 2D and 3D elements.
+ *
+ *  Configurable with the following options in the passed ParameterStore:
+ *  * MapParticlesHost/tol: Tolerance to apply when determining if a
+ *    particle is within a geometry object (default 0.0).
  */
-class MapParticles2D {
+class MapParticlesHost {
 protected:
   /// Disable (implicit) copies.
-  MapParticles2D(const MapParticles2D &st) = delete;
+  MapParticlesHost(const MapParticlesHost &st) = delete;
   /// Disable (implicit) copies.
-  MapParticles2D &operator=(MapParticles2D const &a) = delete;
+  MapParticlesHost &operator=(MapParticlesHost const &a) = delete;
+
+  /// Tolerance to pass to Nektar++ for mapping.
+  REAL tol;
 
   SYCLTargetSharedPtr sycl_target;
   ParticleMeshInterfaceSharedPtr particle_mesh_interface;
-
-  std::unique_ptr<MapParticlesCommon> map_particles_common;
-  std::unique_ptr<MapParticles2DRegular> map_particles_2d_regular;
-  std::unique_ptr<Newton::MapParticlesNewton<Newton::MappingQuadLinear2D>>
-      map_particles_newton_linear_quad;
-
-  int count_regular = 0;
-  int count_deformed = 0;
 
 public:
   /**
@@ -54,7 +49,7 @@ public:
    *  @param particle_mesh_interface ParticleMeshInterface containing 2D
    * Nektar++ cells.
    */
-  MapParticles2D(
+  MapParticlesHost(
       SYCLTargetSharedPtr sycl_target,
       ParticleMeshInterfaceSharedPtr particle_mesh_interface,
       ParameterStoreSharedPtr config = std::make_shared<ParameterStore>());
