+++ include/nektar_interface/particle_cell_mapping/particle_cell_mapping_common.hpp
@@ -8,6 +8,8 @@
 #include <vector>
 
 #include "../particle_mesh_interface.hpp"
+#include "map_particles_common.hpp"
+#include "nektar_interface/parameter_store.hpp"
 #include <SpatialDomains/MeshGraph.h>
 #include <neso_particles.hpp>
 
@@ -207,319 +209,6 @@ inline Geometry3DSharedPtr get_geometry_3d(MeshGraphSharedPtr graph,
   NESOASSERT(false, "Could not find geom in graph.");
   return nullptr;
 }
-
-class MapParticlesCommon {
-protected:
-  /// Disable (implicit) copies.
-  MapParticlesCommon(const MapParticlesCommon &st) = delete;
-  /// Disable (implicit) copies.
-  MapParticlesCommon &operator=(MapParticlesCommon const &a) = delete;
-
-  SYCLTargetSharedPtr sycl_target;
-  std::unique_ptr<ErrorPropagate> ep;
-
-public:
-  MapParticlesCommon(SYCLTargetSharedPtr sycl_target)
-      : sycl_target(sycl_target),
-        ep(std::make_unique<ErrorPropagate>(sycl_target)) {}
-
-  /**
-   *  Returns true if there are particles that were not binned into cells.
-   */
-  inline bool check_map(ParticleGroup &particle_group, const int map_cell = -1,
-                        const bool final_map = true) {
-
-    // Get kernel pointers to the ParticleDats
-    auto cell_id_dat = particle_group.cell_id_dat;
-    auto k_part_cell_ids = cell_id_dat->cell_dat.device_ptr();
-    auto k_part_mpi_ranks = particle_group.mpi_rank_dat->cell_dat.device_ptr();
-
-    // Get iteration set for particles, two cases single cell case or all cells
-    const int max_cell_occupancy = (map_cell > -1)
-                                       ? cell_id_dat->h_npart_cell[map_cell]
-                                       : cell_id_dat->cell_dat.get_nrow_max();
-
-    const int k_cell_offset = (map_cell > -1) ? map_cell : 0;
-    const std::size_t local_size = 256;
-    const auto div_mod = std::div(max_cell_occupancy, local_size);
-    const int outer_size = div_mod.quot + (div_mod.rem == 0 ? 0 : 1);
-    const std::size_t cell_count =
-        (map_cell > -1)
-            ? 1
-            : static_cast<std::size_t>(cell_id_dat->cell_dat.ncells);
-    sycl::range<2> outer_iterset{local_size * outer_size, cell_count};
-    sycl::range<2> local_iterset{local_size, 1};
-    const auto k_npart_cell = cell_id_dat->d_npart_cell;
-
-    this->ep->reset();
-    auto k_ep = this->ep->device_ptr();
-
-    this->sycl_target->queue
-        .submit([&](sycl::handler &cgh) {
-          cgh.parallel_for<>(
-              sycl::nd_range<2>(outer_iterset, local_iterset),
-              [=](sycl::nd_item<2> idx) {
-                const int cellx = idx.get_global_id(1) + k_cell_offset;
-                const int layerx = idx.get_global_id(0);
-                if (layerx < k_npart_cell[cellx]) {
-                  if (k_part_mpi_ranks[cellx][1][layerx] < 0) {
-
-                    if (final_map) {
-                      // if a geom is not found and there is a non-null global
-                      // MPI rank then this function was called after the global
-                      // move and the lack of a local cell / mpi rank is a fatal
-                      // error.
-                      if (((k_part_mpi_ranks)[cellx][0][layerx] > -1) &&
-                          (k_part_mpi_ranks[cellx][1][layerx] < 0)) {
-                        NESO_KERNEL_ASSERT(false, k_ep);
-                      }
-                    } else {
-                      // This loop was called at an intermediate state to
-                      // determine if there exist particles which are not
-                      // mapped into cells. Hence only the local component of
-                      // the mapping dat is checked.
-                      if (k_part_mpi_ranks[cellx][1][layerx] < 0) {
-                        NESO_KERNEL_ASSERT(false, k_ep);
-                      }
-                    }
-                  }
-                }
-              });
-        })
-        .wait_and_throw();
-
-    if (this->ep->get_flag()) {
-      // If the return flag is true there are particles which were not binned
-      // into cells.
-      return true;
-    } else {
-      // If the return flag is false all particles were binned into cells.
-      return false;
-    }
-  }
-};
-
-/**
- *  Class to map particles into Nektar++ cells on the CPU host. Should work for
- *  all 2D and 3D elements.
- */
-class MapParticlesHost {
-protected:
-  /// Disable (implicit) copies.
-  MapParticlesHost(const MapParticlesHost &st) = delete;
-  /// Disable (implicit) copies.
-  MapParticlesHost &operator=(MapParticlesHost const &a) = delete;
-
-  SYCLTargetSharedPtr sycl_target;
-  ParticleMeshInterfaceSharedPtr particle_mesh_interface;
-
-public:
-  /**
-   *  Constructor for mapping class.
-   *
-   *  @param sycl_target SYCLTarget on which to perform mapping.
-   *  @param particle_mesh_interface ParticleMeshInterface containing 2D
-   * Nektar++ cells.
-   */
-  MapParticlesHost(SYCLTargetSharedPtr sycl_target,
-                   ParticleMeshInterfaceSharedPtr particle_mesh_interface)
-      : sycl_target(sycl_target),
-        particle_mesh_interface(particle_mesh_interface) {}
-
-  /**
-   *  Called internally by NESO-Particles to map positions to Nektar++
-   *  triangles and quads.
-   */
-  inline void map(ParticleGroup &particle_group, const int map_cell = -1,
-                  const double tol = 0.0) {
-
-    ParticleDatSharedPtr<REAL> &position_dat = particle_group.position_dat;
-    ParticleDatSharedPtr<REAL> &ref_position_dat =
-        particle_group[Sym<REAL>("NESO_REFERENCE_POSITIONS")];
-    ParticleDatSharedPtr<INT> &cell_id_dat = particle_group.cell_id_dat;
-    ParticleDatSharedPtr<INT> &mpi_rank_dat = particle_group.mpi_rank_dat;
-
-    auto t0 = profile_timestamp();
-    const int rank = this->sycl_target->comm_pair.rank_parent;
-    const int ndim = this->particle_mesh_interface->ndim;
-    const int ncell = this->particle_mesh_interface->get_cell_count();
-    const int nrow_max = mpi_rank_dat->cell_dat.get_nrow_max();
-    auto graph = this->particle_mesh_interface->graph;
-
-    CellDataT<REAL> particle_positions(sycl_target, nrow_max,
-                                       position_dat->ncomp);
-    CellDataT<REAL> ref_particle_positions(sycl_target, nrow_max,
-                                           ref_position_dat->ncomp);
-    CellDataT<INT> mpi_ranks(sycl_target, nrow_max, mpi_rank_dat->ncomp);
-    CellDataT<INT> cell_ids(sycl_target, nrow_max, cell_id_dat->ncomp);
-
-    EventStack event_stack;
-
-    int cell_start, cell_end;
-    if (map_cell < 0) {
-      cell_start = 0;
-      cell_end = ncell;
-    } else {
-      cell_start = map_cell;
-      cell_end = map_cell + 1;
-    }
-
-    double time_copy_from = 0.0;
-    double time_copy_to = 0.0;
-    double time_map_nektar = 0.0;
-    double time_halo_lookup = 0.0;
-
-    for (int cellx = cell_start; cellx < cell_end; cellx++) {
-
-      auto t0_copy_from = profile_timestamp();
-      position_dat->cell_dat.get_cell_async(cellx, particle_positions,
-                                            event_stack);
-      ref_position_dat->cell_dat.get_cell_async(cellx, ref_particle_positions,
-                                                event_stack);
-      mpi_rank_dat->cell_dat.get_cell_async(cellx, mpi_ranks, event_stack);
-      cell_id_dat->cell_dat.get_cell_async(cellx, cell_ids, event_stack);
-
-      event_stack.wait();
-      time_copy_from += profile_elapsed(t0_copy_from, profile_timestamp());
-
-      const int nrow = mpi_rank_dat->cell_dat.nrow[cellx];
-      Array<OneD, NekDouble> global_coord(3);
-      Array<OneD, NekDouble> local_coord(3);
-      auto point = std::make_shared<PointGeom>(ndim, -1, 0.0, 0.0, 0.0);
-      for (int rowx = 0; rowx < nrow; rowx++) {
-
-        // Is this particle already binned into a cell?
-        if ((mpi_ranks)[1][rowx] < 0) {
-
-          // copy the particle position into a nektar++ point format
-          for (int dimx = 0; dimx < ndim; dimx++) {
-            global_coord[dimx] = particle_positions[dimx][rowx];
-            local_coord[dimx] = ref_particle_positions[dimx][rowx];
-          }
-
-          // update the PointGeom
-          point->UpdatePosition(global_coord[0], global_coord[1],
-                                global_coord[2]);
-
-          auto t0_nektar_lookup = profile_timestamp();
-          // get the elements that could contain the point
-          auto element_ids = graph->GetElementsContainingPoint(point);
-          // test the possible local geometry elements
-
-          bool geom_found = false;
-          // check the original nektar++ geoms
-          for (auto &ex : element_ids) {
-            if (ndim == 2) {
-              Geometry2DSharedPtr geom_2d = graph->GetGeometry2D(ex);
-              geom_found =
-                  contains_point_2d(geom_2d, global_coord, local_coord, tol);
-            } else if (ndim == 3) {
-              Geometry3DSharedPtr geom_3d = get_geometry_3d(graph, ex);
-              geom_found =
-                  contains_point_3d(geom_3d, global_coord, local_coord, tol);
-            }
-            if (geom_found) {
-              (mpi_ranks)[1][rowx] = rank;
-              (cell_ids)[0][rowx] = ex;
-              for (int dimx = 0; dimx < ndim; dimx++) {
-                ref_particle_positions[dimx][rowx] = local_coord[dimx];
-              }
-              break;
-            }
-          }
-          time_map_nektar +=
-              profile_elapsed(t0_nektar_lookup, profile_timestamp());
-
-          auto t0_halo_lookup = profile_timestamp();
-          // containing geom not found in the set of owned geoms, now check the
-          // remote geoms
-
-          if (ndim == 2) {
-            if (!geom_found) {
-              for (auto &remote_geom :
-                   this->particle_mesh_interface->remote_triangles) {
-
-                geom_found = contains_point_2d(remote_geom->geom, global_coord,
-                                               local_coord, tol);
-                if (geom_found) {
-                  (mpi_ranks)[1][rowx] = remote_geom->rank;
-                  (cell_ids)[0][rowx] = remote_geom->id;
-                  for (int dimx = 0; dimx < ndim; dimx++) {
-                    ref_particle_positions[dimx][rowx] = local_coord[dimx];
-                  }
-                  break;
-                }
-              }
-            }
-            if (!geom_found) {
-              for (auto &remote_geom :
-                   this->particle_mesh_interface->remote_quads) {
-
-                geom_found = contains_point_2d(remote_geom->geom, global_coord,
-                                               local_coord, tol);
-                if (geom_found) {
-                  (mpi_ranks)[1][rowx] = remote_geom->rank;
-                  (cell_ids)[0][rowx] = remote_geom->id;
-                  for (int dimx = 0; dimx < ndim; dimx++) {
-                    ref_particle_positions[dimx][rowx] = local_coord[dimx];
-                  }
-                  break;
-                }
-              }
-            }
-          } else if (ndim == 3) {
-            if (!geom_found) {
-              for (auto &remote_geom :
-                   this->particle_mesh_interface->remote_geoms_3d) {
-
-                geom_found = contains_point_3d(remote_geom->geom, global_coord,
-                                               local_coord, tol);
-                if (geom_found) {
-                  (mpi_ranks)[1][rowx] = remote_geom->rank;
-                  (cell_ids)[0][rowx] = remote_geom->id;
-                  for (int dimx = 0; dimx < ndim; dimx++) {
-                    ref_particle_positions[dimx][rowx] = local_coord[dimx];
-                  }
-                  break;
-                }
-              }
-            }
-          }
-
-          time_halo_lookup +=
-              profile_elapsed(t0_halo_lookup, profile_timestamp());
-
-          // if a geom is not found and there is a non-null global MPI rank then
-          // this function was called after the global move and the lack of a
-          // local cell / mpi rank is a fatal error.
-          if (((mpi_ranks)[0][rowx] > -1) && !geom_found) {
-            NESOASSERT(false, "No local geometry found for particle");
-          }
-        }
-      }
-
-      auto t0_copy_to = profile_timestamp();
-      ref_position_dat->cell_dat.set_cell_async(cellx, ref_particle_positions,
-                                                event_stack);
-      mpi_rank_dat->cell_dat.set_cell_async(cellx, mpi_ranks, event_stack);
-      cell_id_dat->cell_dat.set_cell_async(cellx, cell_ids, event_stack);
-      event_stack.wait();
-
-      time_copy_to = profile_elapsed(t0_copy_to, profile_timestamp());
-    }
-
-    sycl_target->profile_map.inc("NektarGraphLocalMapperT", "copy_to", 0,
-                                 time_copy_to);
-    sycl_target->profile_map.inc("NektarGraphLocalMapperT", "map_halo", 0,
-                                 time_halo_lookup);
-    sycl_target->profile_map.inc("NektarGraphLocalMapperT", "map_nektar", 0,
-                                 time_map_nektar);
-    sycl_target->profile_map.inc("NektarGraphLocalMapperT", "copy_from", 0,
-                                 time_copy_from);
-    sycl_target->profile_map.inc("NektarGraphLocalMapperT", "map", 1,
-                                 profile_elapsed(t0, profile_timestamp()));
-  }
-};
 
 /**
  *  Count the number of Regular and Deformed geometry objects in a container.
