+++ src/nektar_interface/particle_cell_mapping/map_particles_3d_regular.cpp
@@ -3,9 +3,12 @@
 namespace NESO {
 MapParticles3DRegular::MapParticles3DRegular(
     SYCLTargetSharedPtr sycl_target,
-    ParticleMeshInterfaceSharedPtr particle_mesh_interface)
+    ParticleMeshInterfaceSharedPtr particle_mesh_interface,
+    ParameterStoreSharedPtr config)
     : CoarseMappersBase(sycl_target),
       particle_mesh_interface(particle_mesh_interface) {
+
+  this->tol = config->get("MapParticles3DRegular/tol", 0.0);
 
   // filter out the non-regular elements
   // process locally owned elements
@@ -105,7 +108,7 @@ MapParticles3DRegular::MapParticles3DRegular(
 }
 
 void MapParticles3DRegular::map(ParticleGroup &particle_group,
-                                const int map_cell, const double tol) {
+                                const int map_cell) {
 
   // This method will only map into regular geoms.
   if (this->num_regular_geoms == 0) {
@@ -128,7 +131,7 @@ void MapParticles3DRegular::map(ParticleGroup &particle_group,
   const auto k_map = clm->dh_map->d_buffer.ptr;
   const auto k_map_sizes = clm->dh_map_sizes->d_buffer.ptr;
   const auto k_map_stride = clm->map_stride;
-  const double k_tol = tol;
+  const double k_tol = this->tol;
 
   // Get kernel pointers to the ParticleDats
   const auto position_dat = particle_group.position_dat;
